# ff-split/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinFlow Split Service API
  description: |
    API для сервиса управления совместными расходами в FinFlow.
    
    Сервис предоставляет функциональность для:
    - Управления мероприятиями и событиями
    - Создания и отслеживания транзакций
    - Разделения расходов между пользователями
    - Управления задачами и активностями
    - Работы с категориями и иконками
    
    ## Аутентификация
    
    Защищенные endpoints требуют аутентификации через Bearer токен:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: FinFlow Team
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/event/{id_event}/transaction:
    get:
      tags:
        - transactions
      summary: Получить транзакции мероприятия
      description: Возвращает список транзакций мероприятия
      operationId: getTransactionsByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список транзакций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - transactions
      summary: Создать транзакцию
      description: Создает новую транзакцию в мероприятии
      operationId: createTransaction
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Транзакция создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/transaction/{id_transaction}:
    get:
      tags:
        - transactions
      summary: Получить транзакцию
      description: Возвращает информацию о транзакции
      operationId: getTransactionByID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_transaction
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о транзакции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Транзакция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - transactions
      summary: Обновить транзакцию
      description: Обновляет информацию о транзакции
      operationId: updateTransaction
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_transaction
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Транзакция обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Транзакция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - transactions
      summary: Удалить транзакцию
      description: Удаляет транзакцию
      operationId: deleteTransaction
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_transaction
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Транзакция удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Транзакция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/event/{id_event}/debts:
    get:
      tags:
        - transactions
      summary: Получить долги мероприятия
      description: Возвращает список долгов мероприятия
      operationId: getDebtsByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список долгов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/optimized-debts:
    get:
      tags:
        - transactions
      summary: Получить оптимизированные долги
      description: Возвращает оптимизированные долги мероприятия
      operationId: getOptimizedDebtsByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список оптимизированных долгов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizedDebtListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - transactions
      summary: Оптимизировать долги
      description: Запускает оптимизацию долгов мероприятия
      operationId: optimizeDebts
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Долги оптимизированы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizedDebtListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/user/{id_user}/optimized-debts:
    get:
      tags:
        - transactions
      summary: Получить оптимизированные долги пользователя
      description: Возвращает оптимизированные долги пользователя в мероприятии
      operationId: getOptimizedDebtsByUserID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_user
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список оптимизированных долгов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizedDebtListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Авторизация через Bearer токен
  schemas:
    ErrorResponse:
      type: object
      required:
        - id
        - error
      properties:
        id:
          type: string
          description: ID запроса
        error:
          $ref: '#/components/schemas/ErrorResponseDetail'
    ErrorResponseDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
          example: "validation"
        message:
          type: string
          description: Сообщение об ошибке
          example: "invalid input"
        data:
          $ref: '#/components/schemas/ErrorResponseAdditionalData'
    ErrorResponseAdditionalData:
      type: object
      properties:
        slug:
          type: string
          description: Дополнительная информация
          example: "slug"
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Статус успеха операции
...
    TransactionRequest:
      type: object
      required:
        - name
        - amount
        - from_user
        - type
        - users
      properties:
        name:
          type: string
          description: Название транзакции
        amount:
          type: number
          format: double
          description: Общая сумма транзакции
        from_user:
          type: integer
          format: int64
          description: Внутренний ID пользователя, который заплатил
        type:
          type: string
          enum: [percent, amount, units]
          description: Тип распределения
        users:
          type: array
          items:
            type: integer
            format: int64
          description: Список ID пользователей-участников
        portion:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Распределение долей
        transaction_category_id:
          type: integer
          description: ID категории транзакции
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID транзакции
        event_id:
          type: integer
          format: int64
          description: ID мероприятия
        name:
          type: string
          description: Название транзакции
        amount:
          type: number
          format: double
          description: Сумма транзакции
        from_user:
          type: integer
          format: int64
          description: Внутренний ID плательщика
        type:
          type: string
          description: Тип распределения
        transaction_category_id:
          type: integer
          description: ID категории транзакции
        datetime:
          type: string
          format: date-time
          description: Дата и время транзакции
        shares:
          type: array
          items:
            $ref: '#/components/schemas/ShareDTO'
          description: Доли пользователей
        debts:
          type: array
          items:
            $ref: '#/components/schemas/DebtDTO'
          description: Долги
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
    ShareDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID доли
        transaction_id:
          type: integer
          description: ID транзакции
        user_id:
          type: integer
          format: int64
          description: Внутренний ID пользователя
        value:
          type: number
          format: double
          description: Размер доли
    DebtDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID долга
        transaction_id:
          type: integer
          description: ID транзакции
        from_user_id:
          type: integer
          format: int64
          description: Внутренний ID должника
        to_user_id:
          type: integer
          format: int64
          description: Внутренний ID кредитора
        amount:
          type: number
          format: double
          description: Размер долга
    DebtListResponse:
      type: object
      properties:
        debts:
          type: array
          items:
            $ref: '#/components/schemas/DebtDTO'
    OptimizedDebtDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID оптимизированного долга
        event_id:
          type: integer
          format: int64
          description: ID мероприятия
        from_user_id:
          type: integer
          format: int64
          description: Внутренний ID должника
        to_user_id:
          type: integer
          format: int64
          description: Внутренний ID кредитора
        amount:
          type: number
          format: double
          description: Размер долга
    OptimizedDebtListResponse:
      type: object
      properties:
        optimized_debts:
          type: array
          items:
            $ref: '#/components/schemas/OptimizedDebtDTO'
...